+++ src/os2datascanner/core_organizational_structure/models/syncedpermission.py
@@ -11,11 +11,18 @@ class SyncedPermission(models.Model):
         abstract = True
         permissions = [
             # SBSYS
-            ("view_sbsys_tab", _("Can view the SBSYS tab in Report")),
+            ("view_sbsys_tab", _("Can view the SBSYS tab in the report module")),
             # Leader Statistics Page
             ("filter_scannerjob_leader_overview",
                 _("Can filter by scannerjob in the leader overview")),
             # Scannerjob list
             ("view_scannerjob_list", _("Can view the scannerjob list in the report module")),
-            ("delete_documentreports", _("Can permanently delete reports in the report module"))
+            ("delete_documentreports", _("Can permanently delete reports in the report module")),
+            # Withheld results
+            ("view_withheld_results",
+                _("Can view the withheld results tab in the report module")),
+            ("handle_withheld_results",
+                _("Can handle withheld results in the report module")),
+            ("distribute_withheld_results",
+                _("Can distribute withheld results to users in the report module"))
         ]